move auth to shardingsphere-common
