+++ shardingsphere-underlying/shardingsphere-common/src/main/java/org/apache/shardingsphere/underlying/common/auth/yaml/config/YamlAuthenticationConfiguration.java
@@ -15,20 +15,21 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sharding.core.yaml.config.common;
+package org.apache.shardingsphere.underlying.common.auth.yaml.config;
 
 import lombok.Getter;
 import lombok.Setter;
 import org.apache.shardingsphere.underlying.common.yaml.config.YamlConfiguration;
 
+import java.util.LinkedHashMap;
+import java.util.Map;
+
 /**
- * Proxy user for YAML.
+ * Authentication configuration for YAML.
  */
 @Getter
 @Setter
-public final class YamlProxyUserConfiguration implements YamlConfiguration {
+public final class YamlAuthenticationConfiguration implements YamlConfiguration {
     
-    private String password;
-    
-    private String authorizedSchemas;
+    private Map<String, YamlProxyUserConfiguration> users = new LinkedHashMap<>();
 }
