+++ shardingsphere-underlying/shardingsphere-common/src/main/java/org/apache/shardingsphere/underlying/common/auth/yaml/swapper/ProxyUserYamlSwapper.java
@@ -15,13 +15,13 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sharding.core.yaml.swapper;
+package org.apache.shardingsphere.underlying.common.auth.yaml.swapper;
 
 import com.google.common.base.Joiner;
 import com.google.common.base.Splitter;
 import com.google.common.base.Strings;
 import org.apache.shardingsphere.underlying.common.auth.ProxyUser;
-import org.apache.shardingsphere.sharding.core.yaml.config.common.YamlProxyUserConfiguration;
+import org.apache.shardingsphere.underlying.common.auth.yaml.config.YamlProxyUserConfiguration;
 import org.apache.shardingsphere.underlying.common.yaml.swapper.YamlSwapper;
 
 import java.util.Collections;
