+++ shardingsphere-underlying/shardingsphere-common/src/test/java/org/apache/shardingsphere/underlying/common/auth/yaml/swapper/AuthenticationYamlSwapperTest.java
@@ -15,12 +15,12 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sharding.core.yaml.swapper;
+package org.apache.shardingsphere.underlying.common.auth.yaml.swapper;
 
 import org.apache.shardingsphere.underlying.common.auth.Authentication;
 import org.apache.shardingsphere.underlying.common.auth.ProxyUser;
-import org.apache.shardingsphere.sharding.core.yaml.config.common.YamlAuthenticationConfiguration;
-import org.apache.shardingsphere.sharding.core.yaml.config.common.YamlProxyUserConfiguration;
+import org.apache.shardingsphere.underlying.common.auth.yaml.config.YamlAuthenticationConfiguration;
+import org.apache.shardingsphere.underlying.common.auth.yaml.config.YamlProxyUserConfiguration;
 import org.junit.Test;
 
 import java.util.Collections;
